pyproject.toml
[project]
name = "send-arh999-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datetime>=5.5",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
